requirements.txt
httpx>=0.27.0
feedparser>=6.0.11
python-dotenv>=1.0.0
apscheduler>=3.10.4
tzdata>=2024.1
